Use float32 compute type for Zoom H6 GPU transcriber

+++ Mute-Voce-2.0-main (1)/Mute-Voce-2.0-main/Mute-Voce-main/zoom_h6_largev3_gpu_transcriber_44100Hz/live_transcribe_zoomh6_gpu.py
@@ -12,7 +12,7 @@ CHANNELS = 1
 CHUNK_DURATION = 5
 DEVICE_NAME_FILTER = "zoom"
 MODEL_SIZE = "large-v3"
-COMPUTE_TYPE = "float16"
+COMPUTE_TYPE = "float32"
 DEVICE_TYPE = "cuda"
 
 # ------------------ INITIALIZATION ------------------
